fix: make the endpoint modification exclusive to the v2 info endpoint
