+++ api/views/mixins.py
@@ -1,12 +1,9 @@
 import uuid
 from base64 import b64encode
-from datetime import timedelta
 from inspect import cleandoc
 from mimetypes import guess_extension, guess_type
 from os import path, remove, statvfs
-from platform import machine
 
-import psutil
 from drf_spectacular.utils import OpenApiParameter, OpenApiTypes, extend_schema
 from hurry.filesize import size
 from rest_framework import status
@@ -24,12 +21,11 @@ from api.serializers.mixins import (
 from celery_tasks import reboot_anthias, shutdown_anthias
 from lib import (
     backup_helper,
-    device_helper,
     diagnostics,
 )
 from lib.auth import authorized
 from lib.github import is_up_to_date
-from lib.utils import connect_to_redis, get_node_mac_address
+from lib.utils import connect_to_redis
 from settings import ZmqPublisher, settings
 
 r = connect_to_redis()
@@ -297,41 +293,6 @@ class AssetsControlViewMixin(APIView):
 
 
 class InfoViewMixin(APIView):
-    def get_anthias_version(self):
-        git_branch = diagnostics.get_git_branch()
-        git_short_hash = diagnostics.get_git_short_hash()
-
-        return '{}@{}'.format(
-            git_branch,
-            git_short_hash,
-        )
-
-    def get_device_model(self):
-        device_model = device_helper.parse_cpu_info().get('model')
-
-        if device_model is None and machine() == 'x86_64':
-            device_model = 'Generic x86_64 Device'
-
-        return device_model
-
-    def get_uptime(self):
-        system_uptime = timedelta(seconds=diagnostics.get_uptime())
-        return {
-            'days': system_uptime.days,
-            'hours': round(system_uptime.seconds / 3600, 2),
-        }
-
-    def get_memory(self):
-        virtual_memory = psutil.virtual_memory()
-        return {
-            'total': virtual_memory.total >> 20,
-            'used': virtual_memory.used >> 20,
-            'free': virtual_memory.free >> 20,
-            'shared': virtual_memory.shared >> 20,
-            'buff': virtual_memory.buffers >> 20,
-            'available': virtual_memory.available >> 20
-        }
-
     @extend_schema(
         summary='Get system information',
         responses={
@@ -342,50 +303,7 @@ class InfoViewMixin(APIView):
                     'loadavg': {'type': 'number'},
                     'free_space': {'type': 'string'},
                     'display_power': {'type': ['string', 'null']},
-                    'up_to_date': {'type': 'boolean'},
-                    'anthias_version': {'type': 'string'},
-                    'device_model': {'type': 'string'},
-                    'uptime': {
-                        'type': 'object',
-                        'properties': {
-                            'days': {'type': 'integer'},
-                            'hours': {'type': 'number'}
-                        }
-                    },
-                    'memory': {
-                        'type': 'object',
-                        'properties': {
-                            'total': {'type': 'integer'},
-                            'used': {'type': 'integer'},
-                            'free': {'type': 'integer'},
-                            'shared': {'type': 'integer'},
-                            'buff': {'type': 'integer'},
-                            'available': {'type': 'integer'}
-                        }
-                    },
-                    'mac_address': {'type': 'string'}
-                },
-                'example': {
-                    'viewlog': 'Not yet implemented',
-                    'loadavg': 0.1,
-                    'free_space': '10G',
-                    'display_power': 'on',
-                    'up_to_date': True,
-                    'anthias_version': 'main@a1b2c3d',
-                    'device_model': 'Generic x86_64 Device',
-                    'uptime': {
-                        'days': 1,
-                        'hours': 12.5
-                    },
-                    'memory': {
-                        'total': 8192,
-                        'used': 4096,
-                        'free': 4096,
-                        'shared': 0,
-                        'buff': 1024,
-                        'available': 7168
-                    },
-                    'mac_address': '00:11:22:33:44:55'
+                    'up_to_date': {'type': 'boolean'}
                 }
             }
         }
@@ -404,10 +322,5 @@ class InfoViewMixin(APIView):
             'loadavg': diagnostics.get_load_avg()['15 min'],
             'free_space': free_space,
             'display_power': display_power,
-            'up_to_date': is_up_to_date(),
-            'anthias_version': self.get_anthias_version(),
-            'device_model': self.get_device_model(),
-            'uptime': self.get_uptime(),
-            'memory': self.get_memory(),
-            'mac_address': get_node_mac_address(),
+            'up_to_date': is_up_to_date()
         })
